+++ src/modules/utils/variables.py
@@ -1,3 +1,6 @@
+# For main.py
+experiment_name = "Data Drift Experiment"
+
 # For connect_api.py
 resource_id = "f1765b54-a209-4718-8d38-a39237f502b3"  # This resource ID is for 2017 onwards.
 output_file = "data/2025_pipe/data_source.json"
